src/main/java/edu/hw3/Task8/BackwardIterator.java
package edu.hw3.Task8;

import java.util.Iterator;
import java.util.List;
import java.util.NoSuchElementException;

public class BackwardIterator<T> implements Iterator<T> {
    private List<T> elements;
    private int position;

    public BackwardIterator(List<T> elements) {
        this.elements = elements;
        this.position = elements.size() - 1;
    }

    @Override
    public boolean hasNext() {
        return position >= 0;
    }

    @Override
    public T next() {
        if (!hasNext()) {
            throw new NoSuchElementException();
        }

        return elements.get(position--);
    }
}
